refactor: supprime code mort et code dupliqué sur la cartographie
