--- src/server/cartographie/app/contrats/CartographieSVGContrat.ts
@@ -1,10 +1,0 @@
-export interface CartographieSVGContrat {
-  svg: {
-    g: {
-      path: {
-        "attr-d": string;
-        "attr-territoire-code": string;
-      }[];
-    };
-  };
-}
