--- src/server/infrastructure/api/trpc/routes/cartographieSVG.ts
@@ -1,11 +1,0 @@
-import { SVGService } from "@/server/cartographie/infrastructure/SVGService";
-import {
-  créerRouteurTRPC,
-  procédureProtégée,
-} from "@/server/infrastructure/api/trpc/trpc";
-
-export const cartographieSVGRouter = créerRouteurTRPC({
-  récupérerCartographieSVG: procédureProtégée.query(() =>
-    SVGService.loadSvgAsJson(),
-  ),
-});
